exported DetailsButtons composable into a proper components file
